move table and list to section
