+++ src/02-section/table/Table.js
@@ -1,19 +1,12 @@
-const aliasList = require('../_lib/aliases');
-const setDefaults = require('../_lib/setDefaults');
-const toJson = require('./toJson');
-const toMarkdown = require('./toMarkdown');
+const setDefaults = require('../../_lib/setDefaults');
 const toHtml = require('./toHtml');
+const toMarkdown = require('./toMarkdown');
 const toLatex = require('./toLatex');
+const toJson = require('./toJson');
+const aliasList = require('../../_lib/aliases');
 const defaults = {};
 
-const toText = (list, options) => {
-  return list.map((s) => {
-    let str = s.text(options);
-    return ' * ' + str;
-  }).join('\n');
-};
-
-const List = function(data) {
+const Table = function(data) {
   Object.defineProperty(this, 'data', {
     enumerable: false,
     value: data
@@ -21,14 +14,14 @@ const List = function(data) {
 };
 
 const methods = {
-  lines() {
-    return this.data;
-  },
   links(n) {
     let links = [];
-    this.lines().forEach((s) => {
-      links = links.concat(s.links());
+    this.data.forEach((r) => {
+      Object.keys(r).forEach((k) => {
+        links = links.concat(r[k].links());
+      });
     });
+    //grab a specific link..
     if (typeof n === 'number') {
       return links[n];
     } else if (typeof n === 'string') { //grab a link like .links('Fortnight')
@@ -38,32 +31,40 @@ const methods = {
     }
     return links;
   },
-  markdown(options) {
-    options = setDefaults(options, defaults);
-    return toMarkdown(this, options);
-  },
-  html(options) {
-    options = setDefaults(options, defaults);
-    return toHtml(this, options);
-  },
-  latex(options) {
-    options = setDefaults(options, defaults);
-    return toLatex(this, options);
+  keyValue(options) {
+    let rows = this.json(options);
+    rows.forEach((row) => {
+      Object.keys(row).forEach((k) => {
+        row[k] = row[k].text;
+      });
+    });
+    return rows;
   },
   json(options) {
     options = setDefaults(options, defaults);
-    return toJson(this, options);
+    return toJson(this.data, options);
+  },
+  html(options) {
+    options = setDefaults(options, defaults);
+    return toHtml(this.data, options);
+  },
+  markdown(options) {
+    options = setDefaults(options, defaults);
+    return toMarkdown(this.data, options);
+  },
+  latex(options) {
+    options = setDefaults(options, defaults);
+    return toLatex(this.data, options);
   },
   text() {
-    return toText(this.data);
+    return '';
   }
 };
-
 Object.keys(methods).forEach((k) => {
-  List.prototype[k] = methods[k];
+  Table.prototype[k] = methods[k];
 });
 //add alises, too
 Object.keys(aliasList).forEach((k) => {
-  List.prototype[k] = methods[aliasList[k]];
+  Table.prototype[k] = methods[aliasList[k]];
 });
-module.exports = List;
+module.exports = Table;
