+++ src/03-paragraph/list/index.js
@@ -1,5 +1,5 @@
 const List = require('./List');
-const parseSentence = require('../04-sentence/').oneSentence;
+const parseSentence = require('../../04-sentence/').oneSentence;
 const list_reg = /^[#\*:;\|]+/;
 const bullet_reg = /^\*+[^:,\|]{4}/;
 const number_reg = /^ ?\#[^:,\|]{4}/;
