+++ src/templates/dates/misc.js
@@ -1,17 +1,16 @@
 const format = require('./_format');
 const months = require('./_months');
-const pipeSplit = require('../_parsers/pipeSplit');
+const parse = require('../_parsers/parse');
 
 const misc = {
 
   'reign': (tmpl) => {
     let order = ['start', 'end'];
-    let obj = pipeSplit(tmpl, order);
+    let obj = parse(tmpl, order);
     return `(r. ${obj.start} – ${obj.end})`;
   },
   'circa': (tmpl) => {
-    let order = ['year'];
-    let obj = pipeSplit(tmpl, order);
+    let obj = parse(tmpl, ['year']);
     return `c. ${obj.year}`;
   },
   //we can't do timezones, so fake this one a little bit
@@ -22,20 +21,19 @@ const misc = {
     return format.toText(obj);
   },
   'monthname': (tmpl) => {
-    let order = ['num'];
-    let obj = pipeSplit(tmpl, order);
+    let obj = parse(tmpl, ['num']);
     return months[obj.num] || '';
   },
   //https://en.wikipedia.org/wiki/Template:OldStyleDate
   oldstyledate: (tmpl) => {
     let order = ['date', 'year'];
-    let obj = pipeSplit(tmpl, order);
+    let obj = parse(tmpl, order);
     let str = obj.date;
     if (obj.year) {
       str += ' ' + obj.year;
     }
     return str;
-  },
+  }
 
 };
 module.exports = misc;
